refactor: extract formatSessionTime as an utility
